Now it's getting a tad more structured.
There's now a %Resto{} struct and its associated type.
Restaurants are now inserted into this struct, which needs the
`carte` hierarchy so the informations about the restaurant are not
at the same level as the menus.
I've written a tinay wrapper around `Enum.find` to fetch the restaurants
by name more easily.